+++ src/patchkernel/surface_kernel.hpp
@@ -41,8 +41,8 @@ public:
         // Types definitions
         typedef double (SurfaceKernel::*eval_f_)(long, int &) const;
 
-        void setSpaceDimension(int dimension);
-        int getSpaceDimension(void) const;
+        void setSpaceDimension(short dimension);
+        short getSpaceDimension(void) const;
 
 	virtual ~SurfaceKernel();
         virtual double evalCellArea(long) const;
@@ -80,12 +80,12 @@ protected:
         int                     m_spaceDim;
 
         SurfaceKernel(bool expert);
-        SurfaceKernel(int patch_dim, int space_dim, bool expert);
-        SurfaceKernel(int id, int patch_dim, int space_dim, bool expert);
+        SurfaceKernel(short patch_dim, short space_dim, bool expert);
+        SurfaceKernel(int id, short patch_dim, short space_dim, bool expert);
 #if BITPIT_ENABLE_MPI==1
         SurfaceKernel(MPI_Comm communicator, std::size_t haloSize, bool expert);
-        SurfaceKernel(int patch_dim, int space_dim, MPI_Comm communicator, std::size_t haloSize, bool expert);
-        SurfaceKernel(int id, int patch_dim, int space_dim, MPI_Comm communicator, std::size_t haloSize, bool expert);
+        SurfaceKernel(short patch_dim, short space_dim, MPI_Comm communicator, std::size_t haloSize, bool expert);
+        SurfaceKernel(int id, short patch_dim, short space_dim, MPI_Comm communicator, std::size_t haloSize, bool expert);
 #endif
         
 };
