+++ test/levelset/test_levelset_00002.cpp
@@ -42,7 +42,7 @@
 */
 int subtest_001()
 {
-    int dimensions(3) ;
+    short dimensions(3) ;
 
     // Input geometry
     std::unique_ptr<bitpit::SurfUnstructured> STL( new bitpit::SurfUnstructured(2, 3) );
